requirements.txt
python-telegram-bot 
pandas
ipykernel
scikit-learn 
torch 
transformers 
datasets
